requirements.txt
jupyter
notebook
pandas
numpy
matplotlib
seaborn
scikit-learn
scikit-image
pillow
pyarrow
kaggle
imagehash
pyyaml
tqdm
joblib
fastapi
uvicorn
python-multipart
torch
torchvision